files/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
prometheus-client==0.6.0
python-bitcoinrpc==1.0
requests==2.21.0
urllib3==1.24.2
cfscrape==2.0.8
